requirements.txt
# Copybook Aware EBCDIC Data Converter
lxml>=6.0.2
pydantic>=2.12.5
pyyaml>=6.0.3
python-multipart>=0.0.22
requests>=2.32.5
